Replace yocton_open() with yocton_read_from().
Move responsibility to open the input file to the caller, and have them
pass a FILE* pointer. This gives more flexibility since a FILE* can be
generated in a variety of different ways (with eg. glibc() it's possible
to construct FILEs that call arbitrary callbacks). It also simplifies
the code.