Add *_bytes APIs, allow arbitrary binary strings.
Clarify our philosophical approach to Unicode -- UTF-8 is recommended
but not required, and strings can contain arbitrary binary data. We add
a bunch of tests to confirm that this really is true. Some internal
refactoring is necessary in order to base strings on the new
`struct yocton_buffer` type that can contain arbitrary byte strings.