Eliminate CHECK_OR_GOTO_FAIL.
I was never entirely comfortable with the use of goto here, even though
I do believe that handling errors like this is the one scenario where use
of a goto is justified. A scan through how I was using this macro found that
in the majority of cases, it could be replaced by an equivalent macro,
`CHECK_OR_RETURN`. In a couple of cases where cleanup was being performed,
goto could be avoided just by adding an extra function, and I believe that
the result is cleaner too.