Eliminate redundant error message.
This is only triggered during an EOF while parsing an escape sequence.
Because read_next_byte already sets an error in the case of an EOF, this
message will never actually be set in practice, and indeed we have no
test for this message. So just do the more basic check instead.