Switch to a new method for testing string parsing.
I want to eliminate the ability to use the \x escapes to represent
non-ASCII characters, in order to make the Python API usable. To do
this though, we need a new way to test the result of decoding strings.
So add a new "output" property that writes the string contents into a
buffer and compares that against output contained in special comments.