Lock down \x escapes to just control characters.
For the Python API the \x escape needs to always represent one character.
In a multi-byte encoding like UTF-8 this means that we must restrict it
to just the ASCII range. After some additional consideration I'm locking
it down even further to just the control character range (0x00-0x1f).
The rationale here is that control characters are the only characters
that are disallowed in bare form inside strings. They are therefore the
only characters for which special escaping is needed (except for \\ and
\").