Rename "fields" to "properties".
This change comes after much careful consideration. Here are some of
the reasons:

1. The most pressing reason for it is that it makes documentation clearer.
In adding the `YOCTON_FIELD_*` macros, having two things both called a "field"
(C struct field / Yocton field) makes things more confusing and harder to
follow.

2. Properties were named "fields" through analogy with C struct fields, but it
is a misleading/false analogy. The key difference is that in Yocton, it is
possible to have multiple properties with the same name. So the name "field"
suggests a uniqueness that is not present.

3. More generally I find that the name "field" implies the same uniqueness
described in (2), though this is somewhat open to interpretation since I cannot
immediately find a definition that confirms this.