Remove YOCTON_FIELD_* macros.
This comes after much deliberation. Here's some justification for why I
am getting rid of these macros:

1. They are unnecessary. These macros just duplicate what is already
possible with the YOCTON_VAR_* macros, and predate the addition of those
macros. As time has gone on, it seems increasingly like these are the
better design.

2. They are less flexible than the YOCTON_VAR_* alternatives. The
YOCTON_FIELD_* macros have the built-in assumption that the property
name matches the struct field name, which is not always the case.

3. They do not add much (if anything) in terms of brevity; I originally
thought it best to keep the YOCTON_FIELD_* macros as more concise versions,
but the two have the same number of arguments, so not much is saved beyond
a few characters.

4. They are less clear than the YOCTON_VAR_* alternatives. Consider
the following:

```
YOCTON_VAR_INT(p, foo, int, x->foo);
YOCTON_FIELD_INT(p, *x, int, foo);
```

Both of these do the same thing, but `x->foo` is much more clearly an
lvalue, while the YOCTON_FIELD_INT version hides the thing being set
behind an implicit `(*x).foo`. There's also a "gotcha" in that it takes
a bit of thought to figure out whether to write `x` or `*x`.

5. They add extra bloat to the library that complicates the documentation.  In
starting to write the parse API guide this has been especially clear since
there are two versions of every macro to document. The recent addition of the
*_ARRAY macros has made the bloat even more obvious. The Doxygen documentation
is more complicated since there are more macros to dig through to find the
right thing to use.