Eliminate stringification of property names.
Now that the YOCTON_FIELD_* macros are gone, this is no longer necessary
(those macros relied on the property/field name being parsed as a plain
symbol name). Having the API use a string value is less magic, makes it
more obvious that this is the property name (since a string is passed to
the actual API functions), and also will allow the parse/write APIs to
be matched more naturally, eg.
```c
  #define PROP_NAME "foobar"

  YOCTON_VAR_STRING(p, PROP_NAME, varname);
  yoctonw_prop(w, PROP_NAME, varname);
```