Add YOCTON_VAR_PTR() for allocating pointers.
This is most likely useful for struct field members that optionally
point to other structs, with the latter structs only being allocated
when necessary. In the test case here we just allocate space for a
single integer, but that's fine. It's important that when a property
is being used to allocate something, we don't get that property twice,
because if we did, the pointer might be overwritten causing a memory
leak. Therefore we require such pointer fields to be initialized to
NULL and check that the value is NULL before setting them.