Rephrase doc comments to be more specific.
"The property currently being parsed" is ambiguous since multiple
properties may currently be being parsed. So instead, be specific:
it is the property provided to the macro that is checked.

As part of this, rename the property argument to always be `property`
instead of `prop` as it makes the doc comment clearer. This also makes
the macros more consistent as the name was not always the same.